# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Business Manager API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales:
    get:
      operationId: getSales
      tags: [sales]
      summary: Get all sales
      parameters:
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a new sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleBody"
      responses:
        "201":
          description: Sale created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get a sale by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateSale
      tags: [sales]
      summary: Update a sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSaleBody"
      responses:
        "200":
          description: Sale updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSale
      tags: [sales]
      summary: Delete a sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /inventory:
    get:
      operationId: getInventory
      tags: [inventory]
      summary: Get all inventory items
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create a new inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemBody"
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryItemBody"
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Sale:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        buyPrice:
          type: number
        sellPrice:
          type: number
        profit:
          type: number
        notes:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        inventoryItemId:
          type: ["integer", "null"]
        saleDate:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - productName
        - quantity
        - buyPrice
        - sellPrice
        - profit
        - images
        - saleDate
        - createdAt
        - updatedAt
    CreateSaleBody:
      type: object
      properties:
        productName:
          type: string
        quantity:
          type: integer
        buyPrice:
          type: number
        sellPrice:
          type: number
        notes:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        inventoryItemId:
          type: ["integer", "null"]
        saleDate:
          type: string
      required:
        - productName
        - quantity
        - buyPrice
        - sellPrice
        - saleDate
    UpdateSaleBody:
      type: object
      properties:
        productName:
          type: string
        quantity:
          type: integer
        buyPrice:
          type: number
        sellPrice:
          type: number
        notes:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        inventoryItemId:
          type: ["integer", "null"]
        saleDate:
          type: string
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        quantity:
          type: integer
        buyPrice:
          type: number
        sellPrice:
          type: number
        description:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - name
        - quantity
        - buyPrice
        - sellPrice
        - images
        - createdAt
        - updatedAt
    CreateInventoryItemBody:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: integer
        buyPrice:
          type: number
        sellPrice:
          type: number
        description:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
      required:
        - name
        - quantity
        - buyPrice
        - sellPrice
    UpdateInventoryItemBody:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: integer
        buyPrice:
          type: number
        sellPrice:
          type: number
        description:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
...