# lib/api-zod/src/generated/types/createInventoryItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Business Manager API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateInventoryItemBody {
  name: string;
  quantity: number;
  buyPrice: number;
  sellPrice: number;
  /** @nullable */
  description?: string | null;
  images?: string[];
}