# lib/api-zod/src/generated/types/createSaleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Business Manager API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSaleBody {
  productName: string;
  quantity: number;
  buyPrice: number;
  sellPrice: number;
  /** @nullable */
  notes?: string | null;
  images?: string[];
  /** @nullable */
  inventoryItemId?: number | null;
  saleDate: string;
}